requirements.txt
affine==2.3.0
arrow==1.2.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
contextily==1.2.0
croniter==1.2.0
cycler==0.11.0
dask==2022.1.1
distributed==2022.1.1
docker==5.0.3
docopt==0.6.2
Fiona==1.8.21
fonttools==4.29.1
fsspec==2022.1.0
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
google-api-core==2.4.0
google-auth==2.5.0
google-cloud-aiplatform==1.10.0
google-cloud-bigquery==2.32.0
google-cloud-core==2.2.2
google-cloud-secret-manager==2.9.0
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
hdfs==2.6.0
HeapDict==1.0.1
idna==3.3
importlib-resources==5.4.0
isort==5.10.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
locket==0.2.1
lxml==4.7.1
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-oneofschema==3.0.1
matplotlib==3.5.1
mercantile==1.2.1
msgpack==1.0.3
munch==2.5.0
mypy-extensions==0.4.3
ndjson==0.3.1
numpy==1.22.2
packaging==21.3
pandas==1.4.0
partd==1.2.0
pendulum==2.1.2
Pillow==9.0.1
pkg_resources==0.0.0
prefect==0.15.13
proto-plus==1.19.9
protobuf==3.19.4
psutil==5.9.0
py4j==0.10.9.3
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
pyproj==3.3.0
pyspark==3.2.1
python-box==5.4.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
rasterio==1.2.10
requests==2.27.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
Shapely==1.8.0
six==1.16.0
snuggs==1.4.7
sortedcontainers==2.4.0
tabulate==0.8.9
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
toolz==0.11.2
tornado==6.1
urllib3==1.26.8
websocket-client==1.2.3
wget==3.2
xmltodict==0.12.0
xyzservices==2022.2.0
zict==2.0.0
zipp==3.7.0
